adk/mcp-example/pyproject.toml
[project]
name = "mcp-adk"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "google-adk>=1.16.0",
    "starlette>=0.46.2",
    "uvicorn>=0.34.2",
    "sse-starlette>=2.3.5",
    "starlette>=0.46.2",
]


